refactor: 페이지네이션 관련 공통 타입 분리 #35
