+++ src/types/models/common.ts
@@ -17,3 +17,24 @@ export interface MarkerState {
   isVisible: boolean;
   isClicked: boolean;
 }
+
+
+export interface LikeDto {
+  likeCount: number;
+  isLiked?: boolean;
+}
+
+export interface Sort {
+  empty: boolean;
+  unsorted: boolean;
+  sorted: boolean;
+}
+
+export interface Pageable {
+  sort: Sort;
+  offset: number;
+  pageSize: number;
+  pageNumber: number;
+  paged: boolean;
+  unpaged: boolean;
+}
